Restructure SDK extensions into optional namespaces
